Relocate workflow intrinsic pollution tests
