+++ tests/integration/workflow/source-boundary-intrinsics.test.ts
@@ -5,9 +5,69 @@ import { VeryfrontError } from "#veryfront/errors";
 import {
   requireWorkflowApiBaseUrl,
   requireWorkflowContentSource,
+  type WorkflowSourceAuthority,
 } from "#veryfront/workflow/source-authority.ts";
+import { requireWorkflowSourceIntegrationPolicy } from "#veryfront/workflow/source-integration-policy.ts";
+import { normalizeSourceIntegrationPolicy } from "#veryfront/integrations/source-policy.ts";
 
 describe("workflow source boundaries with hostile ambient intrinsics", () => {
+  it("does not let descriptor prototype pollution turn accessors into data properties", () => {
+    let getterCalls = 0;
+    const authority = Object.defineProperty(
+      { productionMode: true, environmentName: "Production" },
+      "releaseId",
+      {
+        enumerable: true,
+        get() {
+          getterCalls += 1;
+          return "release-from-getter";
+        },
+      },
+    ) as WorkflowSourceAuthority;
+
+    Object.defineProperty(Object.prototype, "value", {
+      value: "polluted-release",
+      configurable: true,
+    });
+    try {
+      assertThrows(
+        () => requireWorkflowContentSource(authority),
+        VeryfrontError,
+        "must contain only own data properties",
+      );
+      assertEquals(getterCalls, 0);
+    } finally {
+      delete (Object.prototype as { value?: unknown }).value;
+    }
+  });
+
+  it("does not let descriptor prototype pollution supply a policy snapshot", () => {
+    const pollutedPolicy = normalizeSourceIntegrationPolicy(undefined);
+    let getterCalls = 0;
+    const run = Object.defineProperty({ id: "polluted-descriptor" }, "sourceIntegrationPolicy", {
+      enumerable: true,
+      get() {
+        getterCalls++;
+        return normalizeSourceIntegrationPolicy(undefined);
+      },
+    });
+
+    Object.defineProperty(Object.prototype, "value", {
+      value: pollutedPolicy,
+      configurable: true,
+    });
+    try {
+      assertThrows(
+        () => requireWorkflowSourceIntegrationPolicy(run as never),
+        VeryfrontError,
+        "invalid source integration policy snapshot",
+      );
+      assertEquals(getterCalls, 0);
+    } finally {
+      delete (Object.prototype as { value?: unknown }).value;
+    }
+  });
+
   it("does not trust a replaced String.prototype.trim for source IDs", () => {
     const originalTrim = String.prototype.trim;
     const originalCharCodeAt = String.prototype.charCodeAt;
